module7-task1: - рефакторинг компонента sorting
